refactor(cli): split pull command files
Amp-Thread-ID: https://ampcode.com/threads/T-019eb30a-8fc6-72de-90a9-6dae7480134e
