+++ pkg/cmd/commands/onboard.go
@@ -1,10 +1,8 @@
 package commands
 
 import (
-	"encoding/json"
 	"errors"
 	"fmt"
-	"io"
 	"os"
 	"path/filepath"
 	"sort"
@@ -25,13 +23,6 @@ type OnboardCmd struct {
 	DryRun      bool   `help:"Preview files without writing them" name:"dry-run"`
 	Force       bool   `help:"Overwrite existing generated files"`
 	SkipVerify  bool   `help:"Skip plan verification after writing files" name:"skip-verify"`
-}
-
-// PullCmd returns live schema from a source environment without writing files.
-type PullCmd struct {
-	Database    string `short:"d" required:"" help:"Database name from SchemaBot server config"`
-	Environment string `short:"e" required:"" help:"Source environment to pull from"`
-	Type        string `help:"Database type" default:"mysql" enum:"mysql"`
 }
 
 // Run executes the onboard command.
@@ -95,31 +86,6 @@ func (cmd *OnboardCmd) Run(g *Globals) error {
 	fmt.Printf("Onboarding complete for %s from %s.\n", resp.Database, resp.Environment)
 	fmt.Println("Next: open a normal PR with these files. SchemaBot will reconcile other configured environments.")
 	return nil
-}
-
-// Run executes the pull command.
-func (cmd *PullCmd) Run(g *Globals) error {
-	ep, err := resolveEndpoint(g.Endpoint, g.Profile)
-	if err != nil {
-		return err
-	}
-	resp, err := client.CallPullSchemaAPI(ep, cmd.Database, cmd.Type, cmd.Environment)
-	if err != nil {
-		if outputSchemaPullRequestError("Pull", cmd.Database, cmd.Environment, err) {
-			return ErrSilent
-		}
-		return fmt.Errorf("pull schema for database %s environment %s: %w", cmd.Database, cmd.Environment, err)
-	}
-	if err := writePullSchemaResponse(os.Stdout, resp); err != nil {
-		return fmt.Errorf("write pull schema response: %w", err)
-	}
-	return nil
-}
-
-func writePullSchemaResponse(w io.Writer, resp *apitypes.PullSchemaResponse) error {
-	enc := json.NewEncoder(w)
-	enc.SetIndent("", "  ")
-	return enc.Encode(resp)
 }
 
 type onboardWritePlan struct {
